refactor(Package): 패키지 이동
- global -> domian
- chat에서 전부 관리
