+++ src/main/java/com/prgrms/mukvengers/domain/chat/service/ChatService.java
@@ -3,11 +3,10 @@ package com.prgrms.mukvengers.domain.chat.service;
 import org.springframework.data.domain.Pageable;
 
 import com.prgrms.mukvengers.domain.chat.dto.response.ChatsInCrew;
-import com.prgrms.mukvengers.global.common.dto.IdResponse;
-import com.prgrms.mukvengers.global.websocket.ChatMessage;
+import com.prgrms.mukvengers.domain.chat.model.ChatMessage;
 
 public interface ChatService {
-	IdResponse save(ChatMessage chatMessage, Long crewId);
+	void save(ChatMessage chatMessage, Long crewId);
 
 	ChatsInCrew getByCrewId(Long crewId, Pageable pageable);
 }
