+++ src/main/java/com/prgrms/mukvengers/domain/chat/service/ChatServiceImpl.java
@@ -9,11 +9,8 @@ import com.prgrms.mukvengers.domain.chat.dto.response.ChatResponse;
 import com.prgrms.mukvengers.domain.chat.dto.response.ChatsInCrew;
 import com.prgrms.mukvengers.domain.chat.mapper.ChatMapper;
 import com.prgrms.mukvengers.domain.chat.model.Chat;
+import com.prgrms.mukvengers.domain.chat.model.ChatMessage;
 import com.prgrms.mukvengers.domain.chat.repository.ChatRepository;
-import com.prgrms.mukvengers.domain.crewmember.exception.MemberNotFoundException;
-import com.prgrms.mukvengers.domain.crewmember.repository.CrewMemberRepository;
-import com.prgrms.mukvengers.global.common.dto.IdResponse;
-import com.prgrms.mukvengers.global.websocket.ChatMessage;
 
 import lombok.RequiredArgsConstructor;
 
@@ -22,23 +19,14 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ChatServiceImpl implements ChatService {
 
-	private final CrewMemberRepository crewMemberRepository;
 	private final ChatRepository chatRepository;
 	private final ChatMapper chatMapper;
 
 	@Override
 	@Transactional
-	public IdResponse save(ChatMessage chatMessage, Long crewId) {
+	public void save(ChatMessage chatMessage, Long crewId) {
 		Chat chat = chatMapper.toChat(chatMessage, crewId);
-
-		Long userId = chatMessage.userId();
-
-		crewMemberRepository.findCrewMemberByCrewIdAndUserId(crewId, userId)
-			.orElseThrow(() -> new MemberNotFoundException(userId));
-
 		chatRepository.save(chat);
-
-		return new IdResponse(chat.getId());
 	}
 
 	@Override
